requirements.txt
annotated-types==0.7.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
construct==2.10.68
construct-typing==0.5.6
contextvars==2.4
contourpy==1.3.2
cycler==0.12.1
Deprecated==1.2.18
docker-pycreds==0.4.0
dominate==2.9.1
einops==0.8.1
eyepy==0.13.0
filelock==3.13.1
fonttools==4.58.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
h5py==3.13.0
hf-xet==1.1.3
huggingface-hub==0.32.4
humanize==4.12.3
idna==3.10
imagecodecs==2023.9.18
imageio==2.37.0
imageio-ffmpeg==0.6.0
immutables==0.21
importlib_resources==6.5.2
Jinja2==3.1.4
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
natsort==8.4.0
networkx==3.3
nibabel==5.3.2
nptyping==2.5.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oct_converter==0.6.2
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.8
protobuf==6.31.0
psutil==7.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pydicom==3.0.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.28.0
setproctitle==1.3.6
shellingham==1.5.4
simpleitk==2.5.0
six==1.17.0
smmap==5.0.2
sympy==1.13.3
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.15
torch==2.7.0+cu126
torchaudio==2.7.0+cu126
torchio==0.20.8
torchsummary==1.5.1
torchvision==0.22.0+cu126
tornado==6.4.2
tqdm==4.67.1
triton==3.3.0
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
visdom==0.2.4
wandb==0.19.11
websocket-client==1.8.0
wrapt==1.17.2